+++ edbee-lib/edbee/util/simpleprofiler.cpp
@@ -94,7 +94,7 @@ void SimpleProfiler::dumpResults()
 
         qint64 totalDuration = 0;
         int totalCallCount = 0;
-        int totalDurationWitoutChilds = 0;
+        qint64 totalDurationWitoutChilds = 0;
         foreach (ProfilerItem* item, items) {
             totalDuration  += item->duration();
             totalCallCount += item->callCount();
@@ -104,7 +104,7 @@ void SimpleProfiler::dumpResults()
         foreach (ProfilerItem* item, items) {
             double durationPercentage = totalDuration > 0 ? 100.0 * static_cast<double>(item->duration()) / static_cast<double>(totalDuration) : 100;
             double callCountPercentage = totalCallCount > 0 ? 100.0 * item->callCount() / totalCallCount : 100;
-            double durationWithoutChildsPercenage = totalDurationWitoutChilds > 0 ? 100.0 * static_cast<double>(item->durationWithoutChilds()) / totalDurationWitoutChilds : 100;
+            double durationWithoutChildsPercenage = totalDurationWitoutChilds > 0 ? 100.0 * static_cast<double>(item->durationWithoutChilds() / totalDurationWitoutChilds) : 100.0;
 
             QString line = QStringLiteral("%1x(%2%) %3ms(%4%) %5ms(%6%) |  %7:%8 %9")
                .arg(item->callCount(), 8).arg(callCountPercentage, 6, 'f', 2)
